Split repositories types module into submodules
