+++ src/repositories/types/common.rs
@@ -1,0 +1,11 @@
+use serde::Deserialize;
+
+#[derive(Deserialize, Debug)]
+#[serde(untagged)]
+pub enum Script {
+    NameOnly(String),
+    Expanded {
+        name: String,
+        version_specific: bool,
+    },
+}
