+++ src/repositories/types/package_target.rs
@@ -1,0 +1,28 @@
+use std::collections::HashMap;
+
+use serde::Deserialize;
+
+use crate::{installer::types::Dependency, repositories::types::Script, utils::checksum::Checksum};
+
+/// Represents the package target data, containing the download url and installer type.
+#[derive(Deserialize, Debug)]
+pub struct PackageTarget {
+    pub url: String,
+    pub checksum: Checksum,
+
+    #[serde(default)]
+    pub dependencies: Vec<Dependency>,
+
+    #[serde(default)]
+    pub build_dependencies: Vec<Dependency>,
+
+    pub skip_symlinking: Option<bool>,
+
+    #[serde(default)]
+    pub script_args: HashMap<String, String>,
+
+    pub build_script: Option<Script>,
+    pub preinstall_script: Option<Script>,
+    pub postinstall_script: Option<Script>,
+    pub test_script: Option<Script>,
+}
