+++ src/repositories/types/repository.rs
@@ -1,0 +1,9 @@
+use serde::Deserialize;
+
+/// Represents the repository metadata, containing repository information.
+#[derive(Deserialize, Debug)]
+pub struct RepositoryMeta {
+    pub name: String,
+    pub description: String,
+    pub maintainers: Vec<String>,
+}
